+++ CHANGELOG.md
@@ -4,10 +4,13 @@ Notable changes will be documented here
 ## Current Release
 ## [v0.8.0-Beta] - XXXX-XX-XX
 ### Added
-- Added support for pushover notifications
+- Added support for pushover & email notifications
+- Added support for data retention
+- User roles. Now you have admins and non-admins.
 ### Changed
 - Everything is now python
 - Moved SMTP settings to config file
+- Hashview Agent is now packaged with hashview (server) under install directory. Can be downloaded from agents menu as admin
 ### Removed
 - Removed hashview agent from local processing. If you want to run hashview AND crack hashes on the same box run the hashview-agent in a seperate screen/tmux session